backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.1
django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
py4j==0.10.9.9
PyJWT==2.10.1
pyodbc==5.2.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn>=20.1.0